+++ faker/providers/file/__init__.py
@@ -324,7 +324,7 @@ class Provider(BaseProvider):
             raise TypeError("Specified file system is invalid.")
 
         root = fs_rule["root"]
-        seperator = fs_rule["separator"]
+        separator = fs_rule["separator"]
 
         path: str = self.file_name(category, extension)
         for _ in range(0, depth):