+++ tests/providers/test_job.py
@@ -13,6 +13,7 @@ from faker.providers.job.hy_AM import Provider as HyAmJobProvider
 from faker.providers.job.ja_JP import Provider as JaJpJobProvider
 from faker.providers.job.ka_GE import Provider as KaGeJobProvider
 from faker.providers.job.ko_KR import Provider as KoKrJobProvider
+from faker.providers.job.mk_MK import Provider as MkMKJobProvider
 from faker.providers.job.pt_BR import Provider as PtBrJobProvider
 from faker.providers.job.pt_PT import Provider as PtPtJobProvider
 from faker.providers.job.ro_RO import Provider as RoRoJobProvider
@@ -162,6 +163,16 @@ class TestKoKr:
     def test_job(self, faker, num_samples):
         for _ in range(num_samples):
             assert faker.job() in KoKrJobProvider.jobs
+
+
+class TestMkMk:
+    """Test mk_MK job provider methods"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            job = faker.job()
+            assert isinstance(job, str)
+            assert job in MkMKJobProvider.jobs
 
 
 class TestPtBr: